Fix for issue #300: URL-Safe panel path
Depending on the settings, the webserver might not always handle
escaped forward-slashes in the way, which this plugin expects it to.

This specifically results in a 404 when trying to edit a redirect-entry,
unless the ```AllowEncodedSlashes NoDecode``` is set for the Apache
Server. The problem occurs in relation to nginx.

Many hosting solutions does now allow customers to change such
settings for the web-server, and so another solution is required.

So, in order to remedy this problem, we replace forward-slash with the
non-visible ascii-characer "GROUP-SEPARATOR" (Oct: 035, Dec: 29,
Hex: 1D). By using a non-visible chracter we ensure that the id
generation from redirect pattern is always unique.

Note that this fix include changes to two parts of the plug-in
code-base. In this file, and in src/panel/components/Tabs/RedirectsTab.vue

See https://github.com/distantnative/retour-for-kirby/issues/300 for context.